% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fledge.R
\name{add_clutch}
\alias{add_clutch}
\title{Add clutch (eggs) to fledgling model}
\usage{
add_clutch(model, eggs)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}

\item{eggs}{a vector of length \code{p} giving eggs for each place}
}
\description{
Add clutch (eggs) to fledgling model
}
