% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fledge_trace.R
\name{add_clutch.trace}
\alias{add_clutch.trace}
\title{Add eggs to trace fledgling model}
\usage{
\method{add_clutch}{trace}(model, eggs)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}

\item{eggs}{a vector of length \code{p} giving eggs for each place}
}
\description{
This function does nothing as trace models are not affected by
endogenous dynamics.
}
