% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fledge.R
\name{add_eggs}
\alias{add_eggs}
\title{Add eggs to fledgling model}
\usage{
add_eggs(model, eggs)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}

\item{eggs}{a vector of length \code{p} giving eggs for each place}
}
\description{
Add eggs to fledgling model
}
