% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birds_interface.R
\name{compute_B_pop}
\alias{compute_B_pop}
\title{Compute total bird population (\eqn{B_{pop}})}
\usage{
compute_B_pop(model)
}
\arguments{
\item{model}{an object from \link[MicroMoB:make_MicroMoB]{MicroMoB::make_MicroMoB}}
}
\value{
a vector of length \code{p}
}
\description{
Compute the total bird population in each place
This method dispatches on the type of \code{model$bird}
}
