% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birds_SIRS.R
\name{compute_B_pop.SIRS}
\alias{compute_B_pop.SIRS}
\title{Compute total SIRS bird population (\eqn{B_{pop}})}
\usage{
\method{compute_B_pop}{SIRS}(model)
}
\arguments{
\item{model}{an object from \link[MicroMoB:make_MicroMoB]{MicroMoB::make_MicroMoB}}
}
\description{
Compute total SIRS bird population (\eqn{B_{pop}})
}
