% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fledge.R
\name{compute_clutch.null}
\alias{compute_clutch.null}
\title{Compute null egg clutches from birds}
\usage{
\method{compute_clutch}{null}(model, N)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}

\item{N}{total number of birds in each patch}
}
\description{
This is to be used with modeling fledgling birth as a trace.
}
