% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birds_interface.R
\name{compute_clutch}
\alias{compute_clutch}
\title{Compute number of eggs laid from birds for each patch}
\usage{
compute_clutch(model)
}
\arguments{
\item{model}{an object from \link[MicroMoB:make_MicroMoB]{MicroMoB::make_MicroMoB}}
}
\value{
a vector of length \code{p} giving the total number of eggs laid by adult birds in each patch
}
\description{
This method dispatches on the type of \code{model$bird}
}
