% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birds_SIRS.R
\name{compute_clutch.SIRS}
\alias{compute_clutch.SIRS}
\title{Compute SIRS egg clutches}
\usage{
\method{compute_clutch}{SIRS}(model)
}
\arguments{
\item{model}{an object from \link[MicroMoB:make_MicroMoB]{MicroMoB::make_MicroMoB}}
}
\description{
Compute SIRS egg clutches
}
