% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fledge_trace.R
\name{compute_clutch.trace}
\alias{compute_clutch.trace}
\title{Compute null egg clutches from birds}
\usage{
\method{compute_clutch}{trace}(model, N)
}
\arguments{
\item{model}{an object from \link[MicroMoB:make_MicroMoB]{MicroMoB::make_MicroMoB}}

\item{N}{total number of birds in each patch}
}
\description{
This is to be used with modeling fledgling birth as a trace.
}
