% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fledge_interface.R
\name{compute_fledge}
\alias{compute_fledge}
\title{Compute number of new fledglings}
\usage{
compute_fledge(model)
}
\arguments{
\item{model}{an object from \link[MicroMoB:make_MicroMoB]{MicroMoB::make_MicroMoB}}
}
\description{
This method dispatches on the type of \code{model$fledge}
}
