% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fledge.R
\name{compute_fledge.trace}
\alias{compute_fledge.trace}
\title{Compute number of new fledglings from trace}
\usage{
\method{compute_fledge}{trace}(model)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}
}
\description{
Compute number of new fledglings from trace
}
\details{
see \link{compute_fledge.trace_deterministic} and \link{compute_fledge.trace_stochastic}
}
