% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birds_interface.R
\name{compute_PsiB}
\alias{compute_PsiB}
\title{Compute time at risk matrix (\eqn{\Psi})}
\usage{
compute_PsiB(model)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}
}
\value{
a matrix with \code{n} rows and \code{p} columns
}
\description{
Compute the home range (time at risk) matrix for bird populations.
This method dispatches on the type of \code{model$bird}.
}
