% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birds_SIRS.R
\name{compute_PsiB.SIRS}
\alias{compute_PsiB.SIRS}
\title{Compute SIRS bird time at risk matrix (\eqn{\Psi})}
\usage{
\method{compute_PsiB}{SIRS}(model)
}
\arguments{
\item{model}{an object from \link[MicroMoB:make_MicroMoB]{MicroMoB::make_MicroMoB}}
}
\description{
Compute SIRS bird time at risk matrix (\eqn{\Psi})
}
