% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_interface.R
\name{compute_q}
\alias{compute_q}
\title{Compute human blood feeding fraction (\eqn{q})}
\usage{
compute_q(model, W, WB)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}

\item{W}{a vector of length \code{p} giving human availability by patch (\eqn{W})}

\item{WB}{a vector of length \code{p} giving bird availability by patch (\eqn{W_{\delta}})}
}
\value{
a vector of length \code{p}
}
\description{
This method dispatches on the type of \code{model$mosquito}
}
