% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans.R
\name{compute_W}
\alias{compute_W}
\title{Compute available humans}
\usage{
compute_W(model)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}
}
\value{
a vector of length \code{p}
}
\description{
Compute available humans
}
