% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birds_interface.R
\name{compute_WB}
\alias{compute_WB}
\title{Compute available birds (\eqn{W_{B}})}
\usage{
compute_WB(model)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}
}
\value{
a vector of length \code{p}
}
\description{
Compute the bird population in each place weighted by home range
and biting weight. This method dispatches on the type of \code{model$bird}.
}
