% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birds_SIRS.R
\name{compute_WB.SIRS}
\alias{compute_WB.SIRS}
\title{Compute available SIRS bird population (\eqn{W_{B}})}
\usage{
\method{compute_WB}{SIRS}(model)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}
}
\description{
Compute available SIRS bird population (\eqn{W_{B}})
}
