% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_SIR.R
\name{compute_wf.SIR}
\alias{compute_wf.SIR}
\title{Compute human biting weights for SIR model (\eqn{w_{f}})}
\usage{
\method{compute_wf}{SIR}(model)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}
}
\description{
Compute human biting weights for SIR model (\eqn{w_{f}})
}
