% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birds_SIRS.R
\name{compute_wfB.SIRS}
\alias{compute_wfB.SIRS}
\title{Compute bird biting weights of SIRS birds (\eqn{w_{f_{B}}})}
\usage{
\method{compute_wfB}{SIRS}(model)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}
}
\description{
Compute bird biting weights of SIRS birds (\eqn{w_{f_{B}}})
}
