% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans.R
\name{compute_x}
\alias{compute_x}
\title{Compute net infectiousness of humans}
\usage{
compute_x(model)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}
}
\value{
a vector of length \code{n}
}
\description{
Compute net infectiousness of humans
}
