% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birds_interface.R
\name{compute_xB}
\alias{compute_xB}
\title{Compute net infectiousness of birds (\eqn{x_{B}})}
\usage{
compute_xB(model)
}
\arguments{
\item{model}{an object from \link[MicroMoB:make_MicroMoB]{MicroMoB::make_MicroMoB}}
}
\value{
a vector of length \code{p}
}
\description{
This is normally computed as the prevalence of disease in each place
multiplied by the transmission efficiency from birds to mosquitoes.
This method dispatches on the type of \code{model$bird}
}
