% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birds_SIRS.R
\name{compute_xB.SIRS}
\alias{compute_xB.SIRS}
\title{Compute net infectiousness of SIRS birds}
\usage{
\method{compute_xB}{SIRS}(model)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}
}
\description{
Compute net infectiousness of SIRS birds
}
