% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birds.R
\name{setup_birds_SIRS}
\alias{setup_birds_SIRS}
\title{Setup birds with SIRS infection model}
\usage{
setup_birds_SIRS(
  model,
  stochastic,
  fledge_disperse,
  theta,
  SIR,
  wf = NULL,
  b = 0.55,
  c = 0.15,
  gamma = 1/5,
  r = 1/120
)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}

\item{stochastic}{should the model update deterministically or stochastically?}

\item{fledge_disperse}{a dispersal matrix for fledglings}

\item{theta}{a matrix giving time spent in each bird's home range}

\item{wf}{biting weights}
}
\description{
This requires the humans to be set up prior to being called.
}
