% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fledge.R
\name{setup_clutch_null}
\alias{setup_clutch_null}
\title{Setup null model for clutches}
\usage{
setup_clutch_null(model)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}
}
\description{
This requires the birds to be set up prior to being called. The
null model is to be used with trace-based fledgling births (from \link{setup_fledge_trace}).
}
