% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic_trace.R, R/fledge_trace.R
\name{setup_fledge_trace}
\alias{setup_fledge_trace}
\title{Setup aquatic (immature) mosquito model with trace (forced) emergence}
\usage{
setup_fledge_trace(model, trace, stochastic)

setup_fledge_trace(model, trace, stochastic)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}

\item{trace}{either a vector of length equal to \code{p} or a matrix with \code{p} rows
and \code{tmax} columns, vectors will be converted to the appropriate matrix internally.}

\item{stochastic}{should the model update deterministically or stochastically?}

\item{lambda}{either a vector of length equal to \code{p}, a matrix with \code{p} rows
and \code{tmax} columns, or a matrix with \code{p} rows and \code{365} columns}
}
\description{
Emergence is passed as a (possibly time varying) parameter which is
decoupled from the adult mosquito dynamics.

This requires the birds to be set up prior to being called.
}
