% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans.R
\name{setup_humans.SIR}
\alias{setup_humans.SIR}
\title{Setup humans with SIR infection model}
\usage{
setup_humans.SIR(
  model,
  theta,
  wf = NULL,
  H,
  SIR,
  b = 0.55,
  c = 0.15,
  gamma = 1/5
)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}

\item{theta}{a time spent matrix}

\item{wf}{biting weights}

\item{H}{vector of strata population sizes}

\item{SIR}{a matrix giving S, I, R counts for each strata}

\item{b}{transmission efficiency (mosquito to human)}

\item{c}{transmission efficiency (human to mosquito)}

\item{gamma}{rate of recovery}
}
\description{
Setup humans with SIR infection model
}
