% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_RM.R
\name{setup_mosquito_RM}
\alias{setup_mosquito_RM}
\title{Setup generalized Ross-Macdonald mosquito model}
\usage{
setup_mosquito_RM(model, stochastic, f, q, eip, p, psi, M, Y, Z)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}

\item{stochastic}{should the model update deterministically or stochastically?}

\item{f}{the feeding rate}

\item{q}{proportion of bites taken on humans and birds}

\item{eip}{the Extrinsic Incubation Period, may either be a scalar, a vector of
length 365, or a vector of length equal to \code{tmax} in the \code{model} object from \link{make_microWNV}}

\item{p}{daily survival probability, may either be a scalar, a vector of
length 365, or a vector of length equal to \code{tmax} in the \code{model} object from \link{make_microWNV}}

\item{psi}{a mosquito dispersal matrix (rows must sum to 1)}

\item{M}{total mosquito density per patch}

\item{Y}{density of incubating mosquitoes per patch}

\item{Z}{density of infectious mosquitoes per patch}
}
\description{
This requires the birds to be set up prior to being called.
}
