% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic_interface.R
\name{step_aqua}
\alias{step_aqua}
\title{Update aquatic (immature) mosquito populations}
\usage{
step_aqua(model)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}
}
\description{
This method dispatches on the type of \code{model$aqua}
}
