% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birds.R
\name{step_birds.SIRS}
\alias{step_birds.SIRS}
\title{Update SIRS bird population}
\usage{
\method{step_birds}{SIRS}(model)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}
}
\description{
Update SIRS bird population
}
\details{
see \link{step_birds.SIRS_deterministic} and \link{step_birds.SIRS_stochastic}
}
