% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birds_SIRS.R
\name{step_birds.SIRS_deterministic}
\alias{step_birds.SIRS_deterministic}
\title{Update SIRS bird population (deterministic)}
\usage{
\method{step_birds}{SIRS_deterministic}(model)
}
\arguments{
\item{model}{an object from \link[MicroMoB:make_MicroMoB]{MicroMoB::make_MicroMoB}}
}
\description{
Update SIRS bird population (deterministic)
}
