% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birds.R
\name{step_birds.SIRS_stochastic}
\alias{step_birds.SIRS_stochastic}
\title{Update SIRS bird population (stochastic)}
\usage{
\method{step_birds}{SIRS_stochastic}(model)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}
}
\description{
Update SIRS bird population (stochastic)
}
