% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fledge_interface.R
\name{step_fledge}
\alias{step_fledge}
\title{Update fledgling populations}
\usage{
step_fledge(model)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}
}
\description{
Update fledgling populations
}
