% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fledge_trace.R
\name{step_fledge.trace}
\alias{step_fledge.trace}
\title{Update fledgling populations for forced births}
\usage{
\method{step_fledge}{trace}(model)
}
\arguments{
\item{model}{an object from \link[MicroMoB:make_MicroMoB]{MicroMoB::make_MicroMoB}}
}
\description{
This function does nothing as trace models are do not have
endogenous dynamics.
}
