% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birds_interface.R
\name{compute_B_pop}
\alias{compute_B_pop}
\title{Compute total bird population}
\usage{
compute_B_pop(model)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}
}
\value{
a vector of length \code{p}
}
\description{
Compute the total bird population in each place
}
