% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birds_SIRS.R
\name{compute_B_pop.SIRS}
\alias{compute_B_pop.SIRS}
\title{Compute total SIRS bird population}
\usage{
\method{compute_B_pop}{SIRS}(model)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}
}
\description{
Compute total SIRS bird population
}
