% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fledge_interface.R
\name{compute_clutch}
\alias{compute_clutch}
\title{Compute egg clutches from birds}
\usage{
compute_clutch(model, N)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}

\item{N}{total number of birds in each patch}
}
\description{
Compute egg clutches from birds
}
