% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fledge.R
\name{compute_fledge}
\alias{compute_fledge}
\title{Compute number of new fledglings}
\usage{
compute_fledge(model)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}
}
\description{
Compute number of new fledglings
}
