% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fledge_trace.R
\name{compute_fledge.trace_deterministic}
\alias{compute_fledge.trace_deterministic}
\title{Compute number of new fledglings from trace (deterministic)}
\usage{
\method{compute_fledge}{trace_deterministic}(model)
}
\arguments{
\item{model}{an object from \link[MicroMoB:make_MicroMoB]{MicroMoB::make_MicroMoB}}
}
\description{
Return the column of the trace matrix for this day.
}
