% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fledge_trace.R
\name{compute_fledge.trace_stochastic}
\alias{compute_fledge.trace_stochastic}
\title{Compute number of new fledglings from trace (stochastic)}
\usage{
\method{compute_fledge}{trace_stochastic}(model)
}
\arguments{
\item{model}{an object from \link[MicroMoB:make_MicroMoB]{MicroMoB::make_MicroMoB}}
}
\description{
Draw a Poisson distributed number of fledglings with mean parameter
from the column of the trace matrix for this day.
}
