% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_RM.R
\name{compute_q}
\alias{compute_q}
\title{Compute human blood feeding fraction (\eqn{q})}
\usage{
compute_q(model, W, WB)
}
\arguments{
\item{model}{an object from \link[MicroMoB:make_MicroMoB]{MicroMoB::make_MicroMoB}}

\item{W}{a vector of length \code{p} giving human availability by patch (\eqn{W})}

\item{WB}{a vector of length \code{p} giving bird availability by patch (\eqn{W_{B}})}
}
\description{
This function is redefined from \code{MicroMoB} because the computation of
\eqn{q} may differ with modeling of birds.
}
