% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birds_interface.R
\name{compute_xB}
\alias{compute_xB}
\title{Compute net infectiousness of birds}
\usage{
compute_xB(model)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}
}
\value{
a vector of length \code{p}
}
\description{
This is normally computed as the prevalence of disease in each place
multiplied by the transmission efficiency from birds to mosquitoes.
}
