% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birds_SIRS.R
\name{setup_birds_SIRS}
\alias{setup_birds_SIRS}
\title{Setup birds with SIRS infection model}
\usage{
setup_birds_SIRS(
  model,
  stochastic,
  fledge_disperse,
  theta,
  SIR,
  mu,
  wf = NULL,
  b = 0.55,
  c = 0.15,
  gamma = 1/5,
  r = 1/120
)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}

\item{stochastic}{should the model update deterministically or stochastically?}

\item{fledge_disperse}{a dispersal matrix for fledglings; this parameter is part
of the adult model rather than fledgling model because the handoff of responsibility
between the two occurs at the moment a fledgling leaves its nest, which is exactly
the point at which they decide where to disperse as a new adult.}

\item{theta}{a matrix giving time spent in each bird's home range}

\item{SIR}{matrix of initial states for each patch}

\item{mu}{either a scalar, a vector of length \code{tmax}, or a vector of length \code{365}
giving daily mortality rates}

\item{wf}{biting weights, should be a vector of length \code{p} or \code{NULL} to use \code{1} for all places/patches.}

\item{b}{transmission efficiency (mosquitoes to birds)}

\item{c}{transmission efficiency (birds to mosquitoes)}

\item{gamma}{inverse of infectious duration (recovery rate)}

\item{r}{inverse of immune duration (rate of loss of immunity)}
}
\description{
This model interfaces with bloodmeals via the vector \code{model$bird$h},
giving the per-capita force of infection for birds in each patch.
}
