% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fledge_trace.R
\name{setup_fledge_trace}
\alias{setup_fledge_trace}
\title{Setup fledglings with trace (forced) births}
\usage{
setup_fledge_trace(model, trace, stochastic)
}
\arguments{
\item{model}{an object from \link[MicroMoB:make_MicroMoB]{MicroMoB::make_MicroMoB}}

\item{trace}{either a vector of length equal to \code{p}, a matrix with \code{p} rows
and \code{tmax} columns, or a matrix with \code{p} rows and \code{365} columns}

\item{stochastic}{should the model update deterministically or stochastically?}
}
\description{
This requires the birds to be set up prior to being called.
}
