% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birds_interface.R
\name{step_birds}
\alias{step_birds}
\title{Update bird population}
\usage{
step_birds(model)
}
\arguments{
\item{model}{an object from \link{make_microWNV}}
}
\description{
This method dispatches on the type of \code{model$bird}.
}
